money column fixes

+++ src/main/java/marketplace/LoadDatabase.java
@@ -1,5 +1,7 @@
 package marketplace;
 
+import java.math.BigDecimal;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
@@ -17,15 +19,15 @@ class LoadDatabase {
 
     @Bean
     CommandLineRunner initDatabase(CustomerRepository customerRepository, ProductRepository productRepository) {
-        Customer rickCustomer = new Customer("Riki Cooper", 300);
+        Customer rickCustomer = new Customer("Riki Cooper", BigDecimal.valueOf(300));
         
     	return args -> {
-            log.info("Preloading " + customerRepository.save(new Customer("Jon Jones", 100)));
+            log.info("Preloading " + customerRepository.save(new Customer("Jon Jones", BigDecimal.valueOf(100))));
             log.info("Preloading " + customerRepository.save(rickCustomer));
 
-            log.info("Preloading " + productRepository.save(new Product("Toy", 150, rickCustomer)));
-            log.info("Preloading " + productRepository.save(new Product("Pen", 50)));
-            log.info("Preloading " + productRepository.save(new Product("TV", 500)));
+            log.info("Preloading " + productRepository.save(new Product("Toy", BigDecimal.valueOf(150), rickCustomer)));
+            log.info("Preloading " + productRepository.save(new Product("Pen", BigDecimal.valueOf(50))));
+            log.info("Preloading " + productRepository.save(new Product("TV", BigDecimal.valueOf(500))));
         };
     }
 }
